Update chocolateyUninstall.ps1
fixed so system doesn't throw a wobbler if it doesn't find a item

+++ automatic/freecad/tools/chocolateyUninstall.ps1
@@ -28,5 +28,8 @@ if ( $packageArgs.fileType -match 'exe' ) {
    $key | ForEach-Object { Write-Warning "- $($_.DisplayName)" }
   }
 } else {
-  Remove-Item -Path $pp.ShortcutFilePath
+  if ($pp.ShortcutFilePath) {
+   Remove-Item -Path $pp.ShortcutFilePath
+  }
+  Write-Information "All Files Removed"
 }
